FortiZTP/requirements.txt
# FortiZTP Python SDK Requirements
# Install with: pip install -r requirements.txt

requests>=2.28.0
PyYAML>=6.0
